requirements.txt
Babel==2.3.4
argparse==1.4.0
astroid==1.4.8
debtcollector==1.8.0
distribute==0.7.3
funcsigs==1.0.2
futures==3.0.5
iso8601==0.1.11
isort==4.2.5
keystoneauth1==2.11.1
lazy-object-proxy==1.2.2
mccabe==0.5.2
monotonic==1.2
msgpack-python==0.4.8
netaddr==0.7.18
netifaces==0.10.4
oslo.config==3.15.0
oslo.i18n==3.8.0
oslo.serialization==2.13.0
oslo.utils==3.16.0
pbr==1.10.0
pep8==1.7.0
positional==1.1.1
pylint==1.6.4
pyparsing==2.1.7
python-keystoneclient==3.4.0
python-swiftclient==3.0.0
pytz==2016.6.1
requests==2.11.0
rfc3986==0.3.1
six==1.10.0
stevedore==1.17.0
wrapt==1.10.8
wsgiref==0.1.2
